mainactivity: extract webview and app coordinators
